--- a/package.json
+++ b/package.json
@@ -67,7 +67,7 @@
     "postcss": "^8.0.9"
   },
   "dependencies": {
-    "@fullhuman/postcss-purgecss": "^4.0.3",
+    "purgecss": "^4.0.3",
     "arg": "^5.0.0",
     "bytes": "^3.0.0",
     "chalk": "^4.1.1",