--- a/package.postcss7.json
+++ b/package.postcss7.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "@fullhuman/postcss-purgecss": "^3.1.3",
+    "purgecss": "^4.0.3",
     "autoprefixer": "^9",
     "postcss": "^7",
     "postcss-functions": "^3",