Hi there!
We decided to use tailwind as a base of our design system. Our codebase is legacy-ish Vue 2 SPA with pretty liberal use of `<style>` blocks in our components, built with webpack (this info will be important later). Tailwind worked great and was delightful to work with. That is, until we got to production build part. 

Turns out, just adding tailwind to our postcss plugins list added a couple of minutes to our build times. Actual usage did not seem to matter: commenting out every `@apply` and `@tailwind` rule did not seem to change anything: build with tailwind plugin included still was couple of minutes slower than without it. Further investigation pointed out at `purge` option as the culprit. Removing it seemed to bring out build times in line with what we expected, but we'd very much like to keep it. Of course, purging unused styles takes time, but minutes still seemed excessive, especially considering that `tailwind-cli` with the same config successfully generates the stylesheet in a couple of seconds.

As it turns out, tailwind plugin runs purgecss on every source stylesheet. Which, in case of legacy-ish code base with a lot of tiny stylesheets, is quite a lot. Sure, if said file does not have `@tailwind` directive, it will get wrapped in `purgecss ignore` block, but extraction of used rules from html and js will still be done from scratch for every one of those tiny stylesheets.

This PR changes the plugin to completely skip `purgecss` in `layers` mode if the source file did not include any tailwind layers. For that, I had to switch from `postcss-purgecss` plugin to calling `purgecss` manually — it does not seem like there is a way to decorate `postcss-purgecss` plugin to call it conditionally for both PostCSS 7 and 8. I guess, the upside of this is that latest purgecss could also be used in compat build. 

Some benchmarks of this PR on our codebase:
Before: `pnpm build  421.86s user 29.51s system 169% cpu 4:25.98 total`
After: `pnpm build  285.95s user 16.38s system 214% cpu 2:21.21 total`

As you can see, it saves 2 minutes for us. It might be also very useful for other people in similar situations: integrating tailwind into existing codebase with modularized css.