@RockSharabShrestha 2021-06-24
Unsubscribe me from this I am not understanding anything.

On Thu, Jun 24, 2021, 15:48 Sergey Tatarintsev ***@***.***>
wrote:

> @adamwathan <https://github.com/adamwathan> is there anything I can do to
> get this PR merged? This is quite an important and important issue for us
> and I'd be happy to do any changes you deem necessary.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/tailwindlabs/tailwindcss/pull/4644#issuecomment-867509421>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AQLGT3767W4IO6XORUMBMX3TUL7GBANCNFSM46VCACMQ>
> .
>