@bradlc 2021-06-25
Hey @SevInf. Comparing this to [PurgeCSS itself](https://github.com/FullHuman/purgecss/blob/bf89ff7179de4782d62d77b0b54b068bcff976cb/packages/purgecss/src/index.ts#L636-L638), are we missing these few lines?

```js
if (this.options.variables) {
  this.variablesStructure.safelist = safelist.variables || [];
}
```