@SevInf 2021-06-28
@bradlc this code is taken from [`purgecss-postcss`](https://github.com/FullHuman/purgecss/blob/master/packages/postcss-purgecss/src/index.ts#L13) plugin and there those lines are also missing. That, however, does not look intentional to me, so I'll add them in a moment.