@SevInf 2021-06-28
Hey @bradlic, so I am back with the fix. You were right, those lines were missing and were breaking `safelist.variables` functionality. This is also something that was already broken when using `purgecss-postcss`. Along with the fix I've also added 2 tests: one for purging the variables and another for respecting safelist for variables. You can see that the latter one also fails on master branch.